paymentconfirmation images , dashbard page and widgets have been made
