--- travelerdubai/lib/bookings/bookings_controller.dart
@@ -1,5 +1,0 @@
-import 'package:get/get.dart';
-
-class BookingController extends GetxController {
-  var status = 'failed'.obs;
-}
